add stream method

+++ src/main/java/com/internet/shop/model/Product.java
@@ -1,15 +1,13 @@
 package com.internet.shop.model;
 
-import java.math.BigDecimal;
-
 public class Product {
     private Long id;
     private String name;
-    private BigDecimal price;
+    private double price;
 
     public Product(String name, double price) {
         this.name = name;
-        this.price = BigDecimal.valueOf(price);
+        this.price = price;
     }
 
     public Long getId() {
@@ -28,11 +26,11 @@ public class Product {
         this.name = name;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
